+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/stack-backend
 
+## 2.5.37
+
+### Patch Changes
+
+- client side account deletion; new account setting style;
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.37
+  - @stackframe/stack-emails@2.5.37
+
 ## 2.5.36
 
 ### Patch Changes